fix(netlify, aws): omit cookies from v1 response (#834)
* fix(aws-lambda): omit cookies from v1 response

* feat: use overloads to improve type safety

* chore: lint

* fix: netlify builder function should share netlify lambda implementation

* fix: don't strip `set-cookie` with v1 payload

* fix(netlify): don't strip `set-cookie` headers
