+++ src/runtime/entries/netlify-builder.ts
@@ -1,4 +1,4 @@
 import { builder } from "@netlify/functions";
-import { handler as _handler } from "#internal/nitro/entries/aws-lambda";
+import { lambda } from "./netlify-lambda";
 
-export const handler = builder(_handler);
+export const handler = builder(lambda);
