+++ src/runtime/entries/netlify.ts
@@ -1,14 +1,8 @@
 import "#internal/nitro/virtual/polyfill";
-import type {
-  Handler,
-  HandlerResponse,
-  HandlerContext,
-  HandlerEvent,
-} from "@netlify/functions/dist/main";
-import type { APIGatewayProxyEventHeaders } from "aws-lambda";
+import type { Handler } from "@netlify/functions/dist/main";
 import { withQuery } from "ufo";
-import { nitroApp } from "../app";
 import { getRouteRulesForPath } from "../route-rules";
+import { lambda } from "./netlify-lambda";
 
 export const handler: Handler = async function handler(event, context) {
   const query = {
@@ -33,50 +27,3 @@ export const handler: Handler = async function handler(event, context) {
 
   return lambda(event, context);
 };
-
-async function lambda(
-  event: HandlerEvent,
-  context: HandlerContext
-): Promise<HandlerResponse> {
-  const query = {
-    ...event.queryStringParameters,
-    ...event.multiValueQueryStringParameters,
-  };
-  const url = withQuery(event.path, query);
-  const method = event.httpMethod || "get";
-
-  const r = await nitroApp.localCall({
-    event,
-    url,
-    context,
-    headers: normalizeIncomingHeaders(event.headers),
-    method,
-    query,
-    body: event.body, // TODO: handle event.isBase64Encoded
-  });
-
-  return {
-    statusCode: r.status,
-    headers: normalizeOutgoingHeaders(r.headers),
-    body: r.body.toString(),
-  };
-}
-
-function normalizeIncomingHeaders(headers?: APIGatewayProxyEventHeaders) {
-  return Object.fromEntries(
-    Object.entries(headers || {}).map(([key, value]) => [
-      key.toLowerCase(),
-      value!,
-    ])
-  );
-}
-
-function normalizeOutgoingHeaders(
-  headers: Record<string, string | string[] | undefined>
-) {
-  return Object.fromEntries(
-    Object.entries(headers)
-      .filter(([key]) => !["set-cookie"].includes(key))
-      .map(([k, v]) => [k, Array.isArray(v) ? v.join(",") : v!])
-  );
-}
